Split out requestActiveTheme
